pyproject.toml
[tool.poetry]
name = "django-jugemaj"
packages = [{ include = 'jugemaj' }]
version = "2.0.1"
description = "A Django app for the Majority Judgment voting system"
authors = ["Guilhem Saurel <guilhem.saurel@laas.fr>"]
license = "BSD-2-Clause"
readme = "README.md"
homepage = "https://github.com/nim65s/django-jugemaj"
documentation = "https://django-jugemaj.readthedocs.io"
classifiers = [
    'Environment :: Web Environment',
    'Framework :: Django',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: BSD License',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Topic :: Internet :: WWW/HTTP',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
]


[tool.poetry.dependencies]
python = "^3.8"
ndh = "^5.1.2"
requests = {version = "^2.23.0", optional = true}
wikidata = {version = "^0.6.1", optional = true}

[tool.poetry.dev-dependencies]
coverage = "^6.2"
black = "^21.12b0"
django-stubs = "^1.9.0"
flake8 = "^4.0.1"
ipython = "^7.30.1"
mypy = "^0.920"
pre-commit = "^2.16.0"
pydocstyle = "^6.0.0"
recommonmark = "^0.6.0"
safety = "^1.10.3"
sphinx = "^3.0.3"


[tool.poetry.extras]
example = ["wikidata", "requests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
